+++ src/gui/helper.cpp
@@ -7,12 +7,4 @@
 // local
 #include "gui/helper.h"
 
-namespace GUI {
-
-inline QString uint8ToQstring(uint8_t x) { return QString::number(x); }
-
-inline uint8_t QStringToUint8(QString s) {
-  return static_cast<uint8_t>(s.toInt());
-}
-
-}  // namespace GUI
+namespace GUI {}  // namespace GUI
