+++ src/solver/dfpn.cpp
@@ -125,11 +125,6 @@ size_t DFPN::selectChild(Node &node, uint32_t &child_phi, uint32_t &delta_2) {
   return best_child_index;
 }
 
-inline void DFPN::saveProofAndDisproofNumbers(Node &node, uint32_t phi,
-                                              uint32_t delta) {
-  tt[node.id] = std::make_pair(phi, delta);
-}
-
 void DFPN::retrieveProofAndDisproofNumbers(Node &node, uint32_t &phi,
                                            uint32_t &delta) {
   if (tt.find(node.id) != tt.end()) {
